backend/requirements.txt
# Django Core
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0
django-filter>=23.0
django-storages>=1.13.0

# Database
psycopg2-binary>=2.9.0
pymongo>=4.3.0
mongoengine>=0.27.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-oauth-toolkit>=2.2.0
cryptography>=40.0.0

# Task Queue
celery>=5.2.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# File Handling
pillow>=9.5.0
python-magic>=0.4.27

# API Documentation
drf-spectacular>=0.26.0

# Environment
python-decouple>=3.8
python-dotenv>=1.0.0

# Monitoring
django-prometheus>=2.3.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.1.0

# Utils
requests>=2.31.0
python-dateutil>=2.8.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.5.0
